style: calendar structure
