--- src/app/(main)/dashboard/components/voyage-dashboard/Calendar/utils/cn.ts
@@ -1,3 +1,0 @@
-export default function cn(...classes: string[]) {
-  return classes.filter(Boolean).join(" ");
-}
